solved eslint warnings
