+++ client/src/components/CategoryData.ts
@@ -1,0 +1,9 @@
+export const categoryData = [
+    { value: 'pens', label: 'Pens' },
+    { value: 'notebooks', label: 'Notebooks' },
+    { value: 'journals', label: 'Journals' },
+    { value: 'cards', label: 'Cards' },
+    { value: 'calendars', label: 'Calendars' },
+    { value: 'planners', label: 'Planners' },
+    { value: 'desk-accessories', label: 'Desk accessories' },
+  ];
